- extend SessionHelper class by adding updateAsEnded and isSedentary methods - change data type of duration in Session from int to long - extend ARBroadcastReceiver by additional logic which contributes to the final version of Sessions detection

+++ app/src/main/java/sk/tuke/ms/sedentti/model/Session.java
@@ -27,7 +27,7 @@ public class Session {
     @DatabaseField
     private long endTimestamp;
     @DatabaseField
-    private int duration;
+    private long duration;
     @DatabaseField(canBeNull = false)
     private Date date;
     @DatabaseField
@@ -82,7 +82,7 @@ public class Session {
         return duration;
     }
 
-    public void setDuration(int duration) {
+    public void setDuration(long duration) {
         this.duration = duration;
     }
